requirements.txt
streamlit
web3
py-solc-x
python-dotenv
subprocess.run